requirements.txt
Django==2.2.8
django-extensions==2.2.6
django-leaflet==0.26.0
